pyproject.toml
[tool.poetry]
name = "birthday-telegram-bot"
version = "2.0.0-alpha"
description = ""
authors = ["orehzzz <olegzzz749@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
peewee = "^3.17.1"
python-telegram-bot = {extras = ["job-queue"], version = "^20.8"}
pytz = "^2024.1"
psycopg2-binary = "^2.9.9"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
